DEV-2131 Change name of temporary target subfolder
File parts are send to a temporary subfolder in the target directory.
That temporary subfolder is a dotfolder. Apparently, the watchfolder
process picks up a dotfolder and potentially watches the files in that
folder. Change the folder to have the suffix `.part` so that the folder
doesn't get picked up.