DEV-2201 Expand logging when transfer failed
When a transfer failed, depending on the cause, it didn't always log the
file it tried to transfer. As there is no correlation ID in the log
entries, it was difficult to find out which file actually failed.

Always add the transfer message as a param in the error log entry.